tests/wpt/tests/html/semantics/forms/the-selectlist-element/selectlist-value-option.tentative.html

<!DOCTYPE html>
<link rel=author href="mailto:jarhar@chromium.org">
<link rel=help href="https://github.com/openui/open-ui/issues/664">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<selectlist id=selectlist>
  <option id=optone>innertext one</option>
  <option id=opttwo value=valueattribute>innertext two</option>
</selectlist>

<script>
test(() => {
  assert_equals(selectlist.value, 'innertext one',
    'The first option should be selected initially.');
  selectlist.value = 'valueattribute';
  assert_equals(selectlist.value, 'valueattribute',
    'Assigning value should look at the options value, not innertext');
}, 'selectlist.value should reflect option.value');
</script>
